requirements.txt
flask
flask-sqlalchemy
groq
python-dotenv
gunicorn
werkzeug